+++ packages/ref/src/cmip_ref/cli/__init__.py
@@ -12,11 +12,11 @@ from loguru import logger
 
 from cmip_ref import __version__
 from cmip_ref.cli import config, datasets, executions, providers, solve
-from cmip_ref.cli._logging import capture_logging
 from cmip_ref.config import Config
 from cmip_ref.constants import config_filename
 from cmip_ref.database import Database
 from cmip_ref_core import __version__ as __core_version__
+from cmip_ref_core.logging import capture_logging
 
 
 class LogLevel(str, Enum):
