+++ packages/ref/src/cmip_ref/executor/local.py
@@ -6,7 +6,7 @@ from cmip_ref.config import Config
 from cmip_ref.database import Database
 from cmip_ref.executor import handle_execution_result
 from cmip_ref.models import MetricExecutionResult as MetricExecutionResultModel
-from cmip_ref_core.executor import redirect_logs
+from cmip_ref_core.logging import redirect_logs
 from cmip_ref_core.metrics import Metric, MetricExecutionDefinition, MetricExecutionResult
 from cmip_ref_core.providers import MetricsProvider
